+++ backend/main.go
@@ -46,7 +46,6 @@ import (
 	"github.com/highlight-run/highlight/backend/stepfunctions"
 	"github.com/highlight-run/highlight/backend/storage"
 	"github.com/highlight-run/highlight/backend/store"
-	"github.com/highlight-run/highlight/backend/timeseries"
 	"github.com/highlight-run/highlight/backend/util"
 	"github.com/highlight-run/highlight/backend/vercel"
 	"github.com/highlight-run/highlight/backend/worker"
@@ -101,7 +100,7 @@ func init() {
 	runtimeParsed = util.Runtime(*runtimeFlag)
 }
 
-func healthRouter(runtimeFlag util.Runtime, db *gorm.DB, tdb timeseries.DB, rClient *redis.Client, ccClient *clickhouse.Client, queue *kafkaqueue.Queue, batchedQueue *kafkaqueue.Queue) http.HandlerFunc {
+func healthRouter(runtimeFlag util.Runtime, db *gorm.DB, rClient *redis.Client, ccClient *clickhouse.Client, queue *kafkaqueue.Queue, batchedQueue *kafkaqueue.Queue) http.HandlerFunc {
 	// only checks kafka because kafka is the only critical infrastructure needed for public graph to be healthy.
 	topic := kafkaqueue.GetTopic(kafkaqueue.GetTopicOptions{Type: kafkaqueue.TopicTypeDefault})
 	batchedTopic := kafkaqueue.GetTopic(kafkaqueue.GetTopicOptions{Type: kafkaqueue.TopicTypeBatched})
@@ -118,7 +117,7 @@ func healthRouter(runtimeFlag util.Runtime, db *gorm.DB, tdb timeseries.DB, rCli
 			return
 		}
 		if runtimeFlag != util.PublicGraph {
-			if err := enhancedHealthCheck(ctx, db, tdb, rClient, ccClient); err != nil {
+			if err := enhancedHealthCheck(ctx, db, rClient, ccClient); err != nil {
 				log.WithContext(ctx).Error(fmt.Sprintf("failed enhanced health check: %s", err))
 				http.Error(w, fmt.Sprintf("failed enhanced health check: %s", err), 500)
 				return
@@ -131,12 +130,12 @@ func healthRouter(runtimeFlag util.Runtime, db *gorm.DB, tdb timeseries.DB, rCli
 	}
 }
 
-func enhancedHealthCheck(ctx context.Context, db *gorm.DB, tdb timeseries.DB, rClient *redis.Client, ccClient *clickhouse.Client) error {
+func enhancedHealthCheck(ctx context.Context, db *gorm.DB, rClient *redis.Client, ccClient *clickhouse.Client) error {
 	const Timeout = 5 * time.Second
 
-	errors := make(chan error, 5)
+	errors := make(chan error, 3)
 	wg := sync.WaitGroup{}
-	wg.Add(4)
+	wg.Add(3)
 	go func() {
 		defer wg.Done()
 		ctx, cancel := context.WithTimeout(ctx, Timeout)
@@ -145,21 +144,6 @@ func enhancedHealthCheck(ctx context.Context, db *gorm.DB, tdb timeseries.DB, rC
 			msg := fmt.Sprintf("failed to query database: %s", err)
 			log.WithContext(ctx).Error(msg)
 			errors <- e.New(msg)
-		}
-	}()
-	go func() {
-		defer wg.Done()
-		ctx, cancel := context.WithTimeout(ctx, Timeout)
-		defer cancel()
-		// in prod, the bucket contains the project id
-		// bucket := "dev-1"
-		if util.IsDevOrTestEnv() {
-			bucket := "dev-bucket"
-			if _, err := tdb.Query(ctx, fmt.Sprintf(`from(bucket: "%s") |> range(start: -1m)`, bucket)); err != nil {
-				msg := fmt.Sprintf("failed to query influx db: %s", err)
-				log.WithContext(ctx).Error(msg)
-				errors <- e.New(msg)
-			}
 		}
 	}()
 	go func() {
@@ -289,7 +273,6 @@ func main() {
 		}
 	}
 
-	tdb := timeseries.New(ctx)
 	stripeClient := &client.API{}
 	stripeClient.Init(stripeApiKey, nil)
 
@@ -349,7 +332,6 @@ func main() {
 	privateResolver := &private.Resolver{
 		ClearbitClient:         clearbit.NewClient(clearbit.WithAPIKey(os.Getenv("CLEARBIT_API_KEY"))),
 		DB:                     db,
-		TDB:                    tdb,
 		MailClient:             sendgrid.NewSendClient(sendgridKey),
 		StripeClient:           stripeClient,
 		StorageClient:          storageClient,
@@ -385,7 +367,7 @@ func main() {
 		AllowCredentials:       true,
 		AllowedHeaders:         []string{"*"},
 	}).Handler)
-	r.HandleFunc("/health", healthRouter(runtimeParsed, db, tdb, redisClient, clickhouseClient, kafkaProducer, kafkaBatchedProducer))
+	r.HandleFunc("/health", healthRouter(runtimeParsed, db, redisClient, clickhouseClient, kafkaProducer, kafkaBatchedProducer))
 
 	zapierStore := zapier.ZapierResthookStore{
 		DB: db,
@@ -480,7 +462,6 @@ func main() {
 		}
 		publicResolver := &public.Resolver{
 			DB:               db,
-			TDB:              tdb,
 			ProducerQueue:    kafkaProducer,
 			BatchedQueue:     kafkaBatchedProducer,
 			DataSyncQueue:    kafkaDataSyncProducer,
@@ -570,7 +551,6 @@ func main() {
 	if runtimeParsed == util.Worker || runtimeParsed == util.All {
 		publicResolver := &public.Resolver{
 			DB:               db,
-			TDB:              tdb,
 			ProducerQueue:    kafkaProducer,
 			BatchedQueue:     kafkaBatchedProducer,
 			DataSyncQueue:    kafkaDataSyncProducer,