+++ frontend/src/pages/Traces/TracesPage.tsx
@@ -27,9 +27,10 @@ import {
 	useGetTracesQuery,
 } from '@/graph/generated/hooks'
 import {
+	MetricAggregator,
 	SortDirection,
 	Trace,
-	TracesMetricType,
+	TracesMetricColumn,
 } from '@/graph/generated/schemas'
 import { useProjectId } from '@/hooks/useProjectId'
 import LogsHistogram from '@/pages/LogsPage/LogsHistogram/LogsHistogram'
@@ -77,6 +78,7 @@ export const TracesPage: React.FC = () => {
 		useGetTracesMetricsQuery({
 			variables: {
 				project_id: projectId!,
+				column: TracesMetricColumn.Duration,
 				group_by: [],
 				params: {
 					query: serverQuery,
@@ -86,16 +88,17 @@ export const TracesPage: React.FC = () => {
 					},
 				},
 				metric_types: [
-					TracesMetricType.Count,
-					TracesMetricType.P50,
-					TracesMetricType.P90,
+					MetricAggregator.Count,
+					MetricAggregator.Avg,
+					MetricAggregator.P50,
+					MetricAggregator.P90,
 				],
 			},
 			skip: !projectId,
 		})
 
 	const histogramBuckets = metricsData?.traces_metrics.buckets
-		.filter((b) => b.metric_type === TracesMetricType.Count)
+		.filter((b) => b.metric_type === MetricAggregator.Count)
 		.map((b) => ({
 			bucketId: b.bucket_id,
 			counts: [{ level: 'traces', count: b.metric_value }],
@@ -103,25 +106,29 @@ export const TracesPage: React.FC = () => {
 
 	const totalCount = _.sumBy(
 		metricsData?.traces_metrics.buckets.filter(
-			(b) => b.metric_type === TracesMetricType.Count,
+			(b) => b.metric_type === MetricAggregator.Count,
 		),
 		(b) => b.metric_value,
 	)
 
 	const metricsBuckets: {
+		avg: number | undefined
 		p50: number | undefined
 		p90: number | undefined
 	}[] = []
 	for (let i = 0; i < metricsData?.traces_metrics.bucket_count; i++) {
-		metricsBuckets.push({ p50: undefined, p90: undefined })
+		metricsBuckets.push({ avg: undefined, p50: undefined, p90: undefined })
 	}
 
 	metricsData?.traces_metrics.buckets.forEach((b) => {
 		switch (b.metric_type) {
-			case TracesMetricType.P50:
+			case MetricAggregator.Avg:
+				metricsBuckets[b.bucket_id].avg = b.metric_value / 1_000_000
+				break
+			case MetricAggregator.P50:
 				metricsBuckets[b.bucket_id].p50 = b.metric_value / 1_000_000
 				break
-			case TracesMetricType.P90:
+			case MetricAggregator.P90:
 				metricsBuckets[b.bucket_id].p90 = b.metric_value / 1_000_000
 				break
 		}