+++ sdk/highlight-nest/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @highlight-run/nest
 
+## 3.4.0
+
+### Minor Changes
+
+-   7355c73f8: Support highlight request context globally to associate async console logs / errors with highlight sessions.
+
+### Patch Changes
+
+-   d452bb1b6: refactor types for highlight session context
+-   Updated dependencies [7355c73f8]
+-   Updated dependencies [747903c88]
+-   Updated dependencies [d452bb1b6]
+    -   @highlight-run/node@3.6.0
+
 ## 3.3.8
 
 ### Patch Changes